updated the update function
